Set InstanceType for OOM Protection during startup
OOM Protection Infra requires the instance type so that it reports to the right metrics. InstanceType is read from the config `accounting.instance.type`.

This config cannot be set using /CLUSTER/CONFIGS API since it is multi-valued. It can only be set using `-configFileName` in the command line.

Moreover, the instance type is known during startup as the code is initialized from either BaseBrokerStarter or BaseServerStarter. Also it should not be configurable.

Therefore the config parameter has been removed. This is OK as unknown config parameters are ignored. InstanceType is passed as a parameter during initialization.