fallback for bad values for periodic controller task conf
This change improves how the controller conf applies the values for periodic task configurations. Previously, invalid new configuration values were throwing exceptions and not letting the service startup. With this update, we now:
Validate that the configured period strings (like 7200s, 2h) pass the validation check in TimeUtils.isPeriodValid().
Log an error message when an invalid period string is configured, and the fallback to the default deprecated old configuration.
Continue to fallback to default deprecated old configuration values when new configuration is not provided.
These are configurations that are updated in this change:
getRetentionControllerFrequencyInSeconds()
getOfflineSegmentIntervalCheckerFrequencyInSeconds()
getRealtimeSegmentValidationFrequencyInSeconds()
getBrokerResourceValidationFrequencyInSeconds()
getStatusCheckerFrequencyInSeconds()
getStatusCheckerWaitForPushTimeInSeconds()
getTaskMetricsEmitterFrequencyInSeconds()
getSegmentLevelValidationIntervalInSeconds()
getMinionInstancesCleanupTaskFrequencyInSeconds()
getMinionInstancesCleanupTaskMinOfflineTimeBeforeDeletionInSeconds()
getStaleInstancesCleanupTaskFrequencyInSeconds()
getStaleInstancesCleanupTaskInstancesRetentionInSeconds()