Codecov tests (#15560)
Removes the skip bytebuffer tests we introduced more than a year ago to increase our confidence in our alternative to LArray buffers. We have been using our alternative in all JVMs for a while and recently we completely removed LArray buffer dependency, so it doesn't seem we need to test in this mode anymore.

Uploads test results to codecov. This is a new feature in codecov described in https://docs.codecov.com/docs/test-analytics.