AggregateReduceFunctions only when it's beneficial
Change to onMatch() of PinotAggregateReduceFunctionsRule that only
reduces SUM and AVG to SUM0 when the resulting SUM0 would possibly benefits from the
rewrite in AggregateCaseToFilterRule. The beneficial cases in
AggregateCaseToFilterRule are:
```
    // A2: SUM0(CASE WHEN x = 'foo' THEN cnt ELSE 0 END)
    //   => SUM0(cnt) FILTER (x = 'foo')
    // B: SUM0(CASE WHEN x = 'foo' THEN 1 ELSE 0 END)
    //   => COUNT() FILTER (x = 'foo')
```
And corresponding cases involving SUM (or AVG) are:
```
    // A2 with SUM: SUM (CASE WHEN x = 'foo' THEN cnt ELSE 0 END)
    //   => CASE WHEN COUNT(*) = 0 THEN NULL
    //           ELSE SUM(cnt) FILTER (WHERE x='foo')
    //      END
    // B with SUM: SUM (CASE WHEN x = 'foo' THEN 1 ELSE 0 END)
    //   => CASE WHEN COUNT(*) = 0 THEN NULL
    //           ELSE COUNT(*) FILTER (WHERE X='foo')
    //      END
```
In these cases, reducing should be beneficial as the CASE WHEN is
converted to FILTER on per-tuple basis.

To account for this, PinotAggregateReduceFunctionsRule only matches
```
Aggregate
    -> Project
```
pattern now.

The reason why these cases are not directly implemented in
CaseToFilterRule is, first Pinot's null-handling is optional, when
`nullHandlingEnabled=false`, SUM returns 0 when no rows are read, and
when the knob is `true`, SUM returns NULL when no rows are read. These
cases would need to be handled seperatedly. Second, implementing the
reduction in AggregateCaseToFilter requires us to put an additional
Project node on top of the Aggregate, this requires basically an entire rewrite
over the rule. IMO it might be easier to allow
AggregateReduceFunctionsRule do the work when it's needed.