Invalidate from cache the stats of opchains that finish successfully.
We have detected that some large queries may end up consuming too much memory for too long in the stats cache used to return stats on timeout.

A way to solve that is to remove the stats as soon as the opchain finishes successfully. This should keep the cache almost empty (given most of the time opchains don't fail) at the cost of losing the stats of fast stages when other stage timeouts.

This is not the best solution but one that is very easy to implement. We will work in better long term solutions.