Enable dynamic turning on/off optProgram rules
Implemented knobs as queryOption that controls turning on / off rules
per query. Currently supported all rules in optProgram. All rules are
enbaled by default. One can specify disabling rule like:
```
SET plannerRule_skipPinotAggregateReduceFunctionsRule=true;
SET plannerRule_skipAggregateCaseToFilterRule=true;
```
The naming convention here is consistent with the rule name shown on
query console from `EXPLAIN PLAN FOR`, it's just `plannerRule_skip<name
as shown on console>`. The name shown on console is the description of
the rule.

This commit introduced a convention that all rules when instantiated
should be explicitly provided with a description. This is added to all
rules used in optProgram now. The rule description will serve as its
identifier in both `EXPLAIN PLAN FOR` and for turning on and off using
knobs.

Relevant constructors for Pinot rules that passes the description as arg
were added.

All constants introduced
are declared in pinot-spi/src/main/java/org/apache/pinot/spi/utils/CommonConstants.java.

A major change to QueryEnvironment is, the optProgram in the QueryEnvironment is now
constructed per-query (to contain only enabled rules), this is
acceptable as creating HepProgram is not expensive. TraitProgram is
already created per-query before this PR.

Several tests are added to QueryCompilationTest and
QueryPlannerRuleOptionsTest.