Assign description for each rule in BASIC_RULES
Explicitly assigned a constant description to each rule in BASIC_RULES.
This serves as the identifier of this rule in both EXPLAIN_PLAN_FOR
output and its queryOption knob.

Relevant constructors for Pinot rules that passes the description as arg
were added.

Simplified the checking logic to directly check from queryOptions.

Added more tests to QueryPlannerRuleOptionsTest.