[WIP] Improve prune() performance of SinglePartitionColumnSegmentPruner
Previously this class stored a map from `segmentName ->
partitionDetails`. This works but means that during prune() we need to
check to see if the partition is matching for each and every segment.

This change observes that generally the total number of partitions will
be much smaller than the number of segments, so it will most often
improve performance to check for matching partitions once.