Remove Sort Limit from EnrichedJoin for now
The previous (incorrect) sort-limit fuse-in logic is temporarily removed.
Fusing-in sort limit node would change EnrichedJoin's left-side to a pipeline breaker as well, since we need to see to all join output rows before sort-limit can emit. This is not much different than not fusing the sort into the join.
When there's no collation, just offset and limit, sort node above join
can do earlyTerminate, the difference is at-most one block.

Sort Limit pushdown might be able to be handled using Calcite rules, IMO
it's better not hardcode this sort-limit pushdown / duplicate down logic
in the operator itself, rather let the planner decide when to enable and
when to disable this behavior.