[SSE] Remove redundant sorting when broker receives sorted block
Previously, the `SelectionDataTableReducer` is blind to server's output
sorting. It performs redundant sort when merging multiple dataTables.

This PR introduces a metadata of block sorting. When a
SelectionResultBlock is created, the `orderByExpressions` metadata is
supplied if it is sorted. The value of this metadata is simply the
string representation of the orderby expressions.

Then in SelectionOperatorService.java, the broker checks for this
metadata and perform mergeWithOrdering on sorted output. If the server
output is not sorted, the broker sorts it as before.