Support disabling rule by default and enabling it with a queryOption
Added queryOption `usePlannerRules` that works similarly to
`skipPlannerRules`. It takes the comma separated list of rules that are disabled by default that to be used.
It only applies to defaultly disabled rules, which is defined in a
static set:
`CommonConstants.Broker#DEFAULT_DISABLED_RULES`. If a rule is set to be enabled
and disabled by queryOption at the same time, it is disabled.

Added `SortJoinTransposeRule`, `SortJoinCopyRule`, and
`AggregateJoinTransposeRule.EXTENDED` in `BASIC_RULES` to be disabled by default.

Adjusted the skipping logic to use empty set instead of null when no
options are provided to avoid NPEs.