Implement partitioned group-by combine for non-order-by non-trim case
Implemented partitioned group-by combine for non-order-by, non-trim case.

This is implemented as 2-phase combine similar to that of DuckDB. In the first phase per-segment results are radix-partitioned,
In the second phase each worker thread picks up a partition to merge the
results into a single hashtable. Then the result hashtables, which are
still radix-partitioned, are
logically stitched together since there are no key collisions between
them. This enables higher level concurrency by eliminating contention
between worker threads.

To be tested and order-by / trim case to be supported. TODO is to add
detail optimizations later. Essentially this is applicable to the broker
as well, since the combine output is still radix-partitioned.