DRY SequentialSortedGroupByCombineOperator and SortedRecordTable
Rewrite `SequentialSortedGroupByCombineOperator` to also merge using a
`SortedRecords` first and at the end construct a `SortedRecordTable`.

Now the only path to `SortedRecordTable` is to init it with already merged and trimmed records. So its fields are effectively static, with no updates
or trimming. It only does the `finish` jobs. Removed all merging
functionalities into `SortedRecordMerger`